# swagger.json
{"swagger": "2.0", "info": {"title": "Kite Trading API", "description": "API for Zerodha Kite trading operations", "version": "1.0.0"}, "basePath": "/", "schemes": ["http"], "securityDefinitions": {"ApiKeyAuth": {"type": "apiKey", "in": "header", "name": "X-Enctoken"}}, "paths": {"/login": {"post": {"summary": "Login to Kite", "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"type": "object", "properties": {"userid": {"type": "string"}, "password": {"type": "string"}, "twofa": {"type": "string"}}}}], "responses": {"200": {"description": "Login successful"}}}}, "/instruments": {"get": {"summary": "Get instruments", "security": [{"ApiKeyAuth": []}], "parameters": [{"name": "exchange", "in": "query", "type": "string", "required": false}], "responses": {"200": {"description": "List of instruments"}}}}, "/historical-data": {"get": {"summary": "Get historical data", "security": [{"ApiKeyAuth": []}], "parameters": [{"name": "instrument_token", "in": "query", "type": "integer", "required": true}, {"name": "from_date", "in": "query", "type": "string", "required": true}, {"name": "to_date", "in": "query", "type": "string", "required": true}, {"name": "interval", "in": "query", "type": "string", "required": true}], "responses": {"200": {"description": "Historical data"}}}}, "/place-order": {"post": {"summary": "Place a new order", "security": [{"ApiKeyAuth": []}], "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"type": "object", "properties": {"variety": {"type": "string"}, "exchange": {"type": "string"}, "tradingsymbol": {"type": "string"}, "transaction_type": {"type": "string"}, "quantity": {"type": "integer"}, "product": {"type": "string"}, "order_type": {"type": "string"}, "price": {"type": "number"}, "validity": {"type": "string"}, "disclosed_quantity": {"type": "integer"}, "trigger_price": {"type": "number"}, "squareoff": {"type": "number"}, "stoploss": {"type": "number"}, "trailing_stoploss": {"type": "number"}, "tag": {"type": "string"}}}}], "responses": {"200": {"description": "Order placed successfully"}}}}, "/orders": {"get": {"summary": "Get all orders", "security": [{"ApiKeyAuth": []}], "responses": {"200": {"description": "List of all orders"}}}}, "/holdings": {"get": {"summary": "Get holdings", "security": [{"ApiKeyAuth": []}], "responses": {"200": {"description": "List of holdings"}}}}, "/positions": {"get": {"summary": "Get positions", "security": [{"ApiKeyAuth": []}], "responses": {"200": {"description": "List of positions"}}}}, "/profile": {"get": {"summary": "Get user profile", "security": [{"ApiKeyAuth": []}], "responses": {"200": {"description": "User profile information"}}}}, "/margins": {"get": {"summary": "Get user margins", "security": [{"ApiKeyAuth": []}], "responses": {"200": {"description": "User margin information"}}}}, "/modify-order": {"put": {"summary": "Modify an existing order", "security": [{"ApiKeyAuth": []}], "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"type": "object", "properties": {"variety": {"type": "string"}, "order_id": {"type": "string"}, "parent_order_id": {"type": "string"}, "quantity": {"type": "integer"}, "price": {"type": "number"}, "order_type": {"type": "string"}, "trigger_price": {"type": "number"}, "validity": {"type": "string"}, "disclosed_quantity": {"type": "integer"}}}}], "responses": {"200": {"description": "Order modified successfully"}}}}, "/cancel-order": {"delete": {"summary": "Cancel an existing order", "security": [{"ApiKeyAuth": []}], "parameters": [{"name": "body", "in": "body", "required": true, "schema": {"type": "object", "properties": {"variety": {"type": "string"}, "order_id": {"type": "string"}, "parent_order_id": {"type": "string"}}}}], "responses": {"200": {"description": "Order cancelled successfully"}}}}}}